feat(coordinator): tcp sockets
Adds tcp sockets to coordinate interaction between the 2 clients
of the experiment. When each round ends, the a message is sent
to the other client to notify that the call is going to end.
It is specially important to enable the other client to download the
webrtcdump before it is deleted(it is deleted when the
rtcpeerconnection disconnects).